fastboot: mmc: print erased bytes instead of hwpart size
In fb_mmc_erase_mmc_hwpart(), we print the amount of erased bytes
using dev_desc->lba.

This works since we guard against blks != dev_desc->lba.

However, "blks" is semantically a variable since it's the amount
of erased blocks.

Note: this is a cosmetic change preparing for the introduction of block
offset in fb_mmc_erase_hwpart()

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>