fastboot: mmc: add offset support to fb_mmc_boot_ops()
In fb_mmc_boot_ops(), we always flash the hwpartition starting at LBA0.

In some cases, like on the amlogic boards [1], we want to preserve the
first LBAs on the hwpartitions.

Prepare for these by adding an offset argument to fb_mmc_boot_ops().

Note: this should be a harmless change, as we pass 0 as an offset.

[1] https://u-boot.readthedocs.io/en/latest/board/amlogic/khadas-vim3l.html#image-creation
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>