configs: prepare khadas-vim3{l}_android for AOSP support
The Khadas VIM3 and VIM3L board are well supported in AOSP[1].
However, there is no mainline U-Boot support for it.
The U-Boot used in AOSP is based on a vendor tree [2]

Add a special defconfig for each board to allow booting Android.

Note: this is a copy-paste of khadas-vim3{l}_defconfig (commit
2bba78076b03e47967180776a8da74a018186480) and will be updated
in future commits.

[1] https://source.android.com/setup/build/devices#vim3_and_vim3l_boards
[2] https://gitlab.com/baylibre/amlogic/atv/u-boot
Change-Id: I44193607cd09505127e95119597f05188647a73b
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>