HACK: usb: gadget: fastboot: call gadget_release() before reset
In the case of "fastboot reboot fastboot", the host PC expects the USB
link to be dropped before the board is reset.
On VIM3/VIM3L boards which use CONFIG_PSCI_RESET, this is not the case.

Call usb_gadget_release() before resetting the board.

Note: this is marked as HACK since:
- not sure all boards need this
- not sure this is the right place to call gadget_release()
- maybe disabling CONFIG_PSCI_RESET + calling usb_gadget_release() in
reset_cpu() is a better solution?

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>